requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
sentence-transformers
faiss-cpu
mistralai
python-docx
PyPDF2
python-dotenv